+++ src/constants/requestMethod.ts
@@ -2,6 +2,7 @@
  * Request methods.
  */
 enum RequestMethod {
+  Delete = 'DELETE',
   Get = 'GET',
   Post = 'POST',
   Put = 'PUT',